rename the internationalization package
