--- src/main/java/internationalization/package-info.java
@@ -1,5 +1,0 @@
-/**
- * @author mivek
- *
- */
-package internationalization;
