+++ src/main/java/io/github/mivekinternationalization/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * @author mivek
+ *
+ */
+package io.github.mivekinternationalization;
